feat: check shared space
